Browser upload
I will be coding the functionality of the game as to what the user needs to do in order to play the game.

Updated main.js file. Showing my first function for the ON BUTTON click. Once clicked the game will start, counter will activate. On another click the game will turn off, counter will turn off, colours will turn off as well as the interval.