Browser upload
We added the eventListener so that the user can click on the "colour" and if the game is on, the player order we push one on the player order array, then we check if the player was right, then call the one function. Further, if the player has not won, we invoke the clearColor() function to clear the colour after a set an amount of time (3 seconds in our case).