Browser upload
Uploaded the js file with the new function check(). This function checks, if the players order isn`t equal with the real order, by considering the last thing that the player clicked, then this isn`t good. It will be good if the player is getting everything correct.
Also if the player got all of the correct then we call the winGame() function. If the player got something wrong, we flash the colours, we invoke the innerHTML to display NO on the count screen. Also if we are in strict mode and the player gets it wrong, the game starts from the first level.